@@ -2,8 +2,8 @@
 	"name": "ng-loaders",
 	"version": "5.0.2",
 	"peerDependencies": {
-		"@angular/common": ">=16.0.0",
-		"@angular/core": ">=16.0.0"
+		"@angular/common": ">=19.0.0",
+		"@angular/core": ">=19.0.0"
 	},
 	"dependencies": {
 		"tslib": "^2.3.0"
