@@ -17,60 +17,62 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "17.0.2",
-		"@angular/common": "17.0.2",
-		"@angular/compiler": "17.0.2",
-		"@angular/core": "17.0.2",
-		"@angular/forms": "17.0.2",
-		"@angular/platform-browser": "17.0.2",
-		"@angular/platform-browser-dynamic": "17.0.2",
-		"@angular/router": "17.0.2",
-		"@nx/angular": "17.1.2",
+		"@angular/animations": "21.1.3",
+		"@angular/common": "21.1.3",
+		"@angular/compiler": "21.1.3",
+		"@angular/core": "21.1.3",
+		"@angular/forms": "21.1.3",
+		"@angular/platform-browser": "21.1.3",
+		"@angular/platform-browser-dynamic": "21.1.3",
+		"@angular/router": "21.1.3",
+		"@nx/angular": "22.4.5",
 		"rxjs": "~7.8.0",
 		"tslib": "^2.3.0",
-		"zone.js": "0.14.2"
+		"zone.js": "0.16.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "17.0.0",
-		"@angular-devkit/core": "17.0.0",
-		"@angular-devkit/schematics": "17.0.0",
-		"@angular-eslint/eslint-plugin": "17.0.1",
-		"@angular-eslint/eslint-plugin-template": "17.0.1",
-		"@angular-eslint/template-parser": "17.0.1",
-		"@angular/cli": "~16.2.0",
-		"@angular/compiler-cli": "17.0.2",
-		"@angular/language-service": "17.0.2",
-		"@nx/cypress": "17.1.2",
-		"@nx/eslint-plugin": "17.1.2",
-		"@nx/jest": "17.1.2",
-		"@nx/js": "17.1.2",
-		"@nx/web": "17.1.2",
-		"@nx/workspace": "17.1.2",
-		"@schematics/angular": "17.0.0",
-		"@types/jest": "^29.4.0",
-		"@types/node": "16.11.7",
-		"@typescript-eslint/eslint-plugin": "6.11.0",
-		"@typescript-eslint/parser": "6.11.0",
+		"@angular-devkit/build-angular": "21.1.3",
+		"@angular-devkit/core": "21.1.3",
+		"@angular-devkit/schematics": "21.1.3",
+		"@angular-eslint/eslint-plugin": "21.2.0",
+		"@angular-eslint/eslint-plugin-template": "21.2.0",
+		"@angular-eslint/template-parser": "21.2.0",
+		"@angular/cli": "21.1.3",
+		"@angular/compiler-cli": "21.1.3",
+		"@angular/language-service": "21.1.3",
+		"@nx/cypress": "22.4.5",
+		"@nx/eslint": "22.4.5",
+		"@nx/eslint-plugin": "22.4.5",
+		"@nx/jest": "22.4.5",
+		"@nx/js": "22.4.5",
+		"@nx/web": "22.4.5",
+		"@nx/workspace": "22.4.5",
+		"@schematics/angular": "21.1.3",
+		"@types/jest": "30.0.0",
+		"@types/node": "^18.16.9",
+		"@typescript-eslint/eslint-plugin": "^8.54.0",
+		"@typescript-eslint/parser": "^8.54.0",
+		"@typescript-eslint/utils": "^8.54.0",
 		"autoprefixer": "^10.4.0",
-		"cypress": "^13.0.0",
-		"eslint": "8.46.0",
-		"eslint-config-prettier": "9.0.0",
+		"cypress": "15.9.0",
+		"eslint": "8.57.1",
+		"eslint-config-prettier": "10.1.8",
 		"eslint-plugin-cypress": "2.15.1",
-		"jest": "^29.4.1",
-		"jest-environment-jsdom": "^29.4.1",
-		"jest-preset-angular": "13.1.4",
-		"ng-packagr": "17.0.0",
-		"nx": "17.1.2",
+		"jest": "30.2.0",
+		"jest-environment-jsdom": "30.2.0",
+		"jest-preset-angular": "16.0.0",
+		"jest-util": "30.2.0",
+		"ng-packagr": "21.1.0",
+		"nx": "22.4.5",
 		"postcss": "^8.4.5",
 		"postcss-import": "~14.1.0",
 		"postcss-preset-env": "~7.5.0",
 		"postcss-url": "~10.1.3",
 		"prettier": "^2.6.2",
 		"tailwindcss": "^3.0.2",
-		"ts-jest": "^29.1.0",
+		"ts-jest": "29.4.6",
 		"ts-node": "10.9.1",
-		"typescript": "5.2.2",
-		"@nx/eslint": "17.1.2"
+		"typescript": "5.9.3"
 	},
 	"description": "Angular component to add CSS loaders to your application",
 	"author": {
