@@ -1,12 +1,5 @@
-import {
-	NgClass,
-	NgComponentOutlet,
-	NgIf,
-	NgSwitch,
-	NgSwitchCase,
-	NgSwitchDefault,
-} from '@angular/common';
-import { Component, Injector, Input } from '@angular/core';
+import { NgClass, NgComponentOutlet } from '@angular/common';
+import { Component, Injector, inject, input } from '@angular/core';
 import { RotatingCubesComponent } from '../rotating-cubes/rotating-cubes.component';
 import { CubeGridComponent } from '../cube-grid/cube-grid.component';
 import { EllipsisComponent } from '../ellipsis/ellipsis.component';
@@ -15,7 +8,7 @@ import { GrowingCircleComponent } from '../growing-circle/growing-circle.compone
 import { PulsingCirclesComponent } from '../pulsing-circles/pulsing-circles.component';
 import { RotatingDotsComponent } from '../rotating-dots/rotating-dots.component';
 import { RotatingSquareComponent } from '../rotating-square/rotating-square.component';
-import { NG_LOADERS_CONFIG, NgLoadersConfig } from '../injection-token';
+import { NG_LOADERS_CONFIG } from '../injection-token';
 
 export type SpinnerType =
 	| 'cube-grid'
@@ -29,23 +22,18 @@ export type SpinnerType =
 
 @Component({
 	selector: 'ngx-loader',
-	standalone: true,
-	imports: [
-		NgIf,
-		NgClass,
-		NgSwitch,
-		NgSwitchCase,
-		NgSwitchDefault,
-		NgComponentOutlet,
-	],
+	imports: [NgClass, NgComponentOutlet],
 	templateUrl: './loader.component.html',
 	styleUrls: ['./loader.component.scss'],
 })
 export class LoaderComponent {
-	private config: NgLoadersConfig = this._injector.get(NG_LOADERS_CONFIG);
-	@Input() spinnerType: SpinnerType = this.config.spinnerType;
-	@Input() coverContainer = false;
-	@Input() fixed: boolean = false;
+	private config = inject(NG_LOADERS_CONFIG);
+	public injector = inject(Injector);
+
+	spinnerType = input<SpinnerType>(this.config.spinnerType);
+	coverContainer = input(false);
+	fixed = input(false);
+
 	componentsMap = {
 		'cube-grid': CubeGridComponent,
 		ellipsis: EllipsisComponent,
@@ -56,6 +44,4 @@ export class LoaderComponent {
 		'rotating-dots': RotatingDotsComponent,
 		'rotating-square': RotatingSquareComponent,
 	};
-
-	constructor(private _injector: Injector) {}
 }
