@@ -2,5 +2,10 @@ import { defineConfig } from 'cypress';
 import { nxE2EPreset } from '@nx/cypress/plugins/cypress-preset';
 
 export default defineConfig({
-	e2e: nxE2EPreset(__dirname),
+	e2e: {
+		...nxE2EPreset(__dirname),
+		// Please ensure you use `cy.origin()` when navigating between domains and remove this option.
+		// See https://docs.cypress.io/app/references/migration-guide#Changes-to-cyorigin
+		injectDocumentDomain: true,
+	},
 });
