@@ -1,12 +1,11 @@
 import { Component, signal } from '@angular/core';
-import { CommonModule, NgIf } from '@angular/common';
+
 
 @Component({
-	selector: 'al-navbar',
-	standalone: true,
-	imports: [NgIf],
-	templateUrl: './navbar.component.html',
-	styleUrls: ['./navbar.component.scss'],
+    selector: 'al-navbar',
+    imports: [],
+    templateUrl: './navbar.component.html',
+    styleUrls: ['./navbar.component.scss']
 })
 export class NavbarComponent {
 	public menuOpen = signal(false);
