@@ -5,4 +5,6 @@ globalThis.ngJest = {
 		errorOnUnknownProperties: true,
 	},
 };
-import 'jest-preset-angular/setup-jest';
+import { setupZoneTestEnv } from 'jest-preset-angular/setup-env/zone';
+
+setupZoneTestEnv();
