@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-  selector: 'ngx-cube-grid',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './cube-grid.component.html',
-  styleUrls: ['./cube-grid.component.scss'],
+    selector: 'ngx-cube-grid',
+    imports: [],
+    templateUrl: './cube-grid.component.html',
+    styleUrls: ['./cube-grid.component.scss']
 })
 export class CubeGridComponent {}
