@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-  selector: 'ngx-ellipsis',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './ellipsis.component.html',
-  styleUrls: ['./ellipsis.component.scss'],
+    selector: 'ngx-ellipsis',
+    imports: [],
+    templateUrl: './ellipsis.component.html',
+    styleUrls: ['./ellipsis.component.scss']
 })
 export class EllipsisComponent {}
