@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-  selector: 'ngx-growing-circle',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './growing-circle.component.html',
-  styleUrls: ['./growing-circle.component.scss'],
+    selector: 'ngx-growing-circle',
+    imports: [],
+    templateUrl: './growing-circle.component.html',
+    styleUrls: ['./growing-circle.component.scss']
 })
 export class GrowingCircleComponent {}
