@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-  selector: 'ngx-pulsing-circles',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './pulsing-circles.component.html',
-  styleUrls: ['./pulsing-circles.component.scss'],
+    selector: 'ngx-pulsing-circles',
+    imports: [],
+    templateUrl: './pulsing-circles.component.html',
+    styleUrls: ['./pulsing-circles.component.scss']
 })
 export class PulsingCirclesComponent {}
