@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
-  selector: 'ngx-rotating-square',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './rotating-square.component.html',
-  styleUrls: ['./rotating-square.component.scss'],
+    selector: 'ngx-rotating-square',
+    imports: [],
+    templateUrl: './rotating-square.component.html',
+    styleUrls: ['./rotating-square.component.scss']
 })
 export class RotatingSquareComponent {}
